backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
SQLAlchemy==2.0.20
pydantic==1.10.13
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2cffi==2.9.0  # PostgreSQL数据库驱动的纯Python替代方案